Hooks have been moved to the child directory
