+++ src/models/IdType.ts
@@ -1,4 +1,7 @@
 /**
- * Workscpace-unique ID for nodes, edges, networks.
+ * Network-unique ID for nodes, edges, networks.
+ *
+ * TODO: What constraints should be placed on this?
+ *
  */
 export type IdType = string
