+++ src/store/hooks/useNdexNetworkSummary.ts
@@ -1,11 +1,11 @@
 // @ts-expect-error-next-line
 import { NDEx } from '@js4cytoscape/ndex-client'
-import { NdexNetworkSummary } from '../models/NetworkSummaryModel'
-import { IdType } from '../models/IdType'
-import { getNetworkSummariesFromDb, putNetworkSummaryToDb } from './persist/db'
+import { NdexNetworkSummary } from '../../models/NetworkSummaryModel'
+import { IdType } from '../../models/IdType'
+import { getNetworkSummariesFromDb, putNetworkSummaryToDb } from '../persist/db'
 
 // check the local cache for ndex network summaries and fetch from NDEx if not found
-export const getNdexNetworkSummary = async (
+export const useNdexNetworkSummary = async (
   ndexNetworkId: IdType | IdType[],
   url: string,
   accessToken?: string,
